Increase precision of model to float64

+++ tests/test_ddp_individual_parameters.py
@@ -49,7 +49,7 @@ def _test_DistributedDataParallelIndividualParameters(rank: int, world_size: int
 
     # Create a toy model and move it to the proper device.
     # This is our non-parallel baseline.
-    non_parallel_model = model_class().to(device)
+    non_parallel_model = model_class().to(device).double()
 
     # Create a DDP model. Note that the weights of this model should
     # match the non-parallel baseline above.
